+++ src/components/mod.rs
@@ -41,5 +41,6 @@ pub mod tab_view;
 pub mod text_input_dialog;
 pub mod time_graph;
 pub mod time_input;
+pub mod view_creation;
 pub mod window;